Refactored buildDicTable code into own test file
